+++ src/modules/kiosks/components/CustomizeScreen.js
@@ -36,7 +36,7 @@ const SupportEmailToolTip = () => (
 
 const MemberCardToolTip = () => (
   <Popup
-    content="This feature can take a couple of minutes to update on the Kiosk."
+    content="Enabling/disabling the member card might take up to 2 minutes to get updated on the kiosk."
     trigger={<Icon color="yellow" name="info circle" />}
   />
 );
@@ -49,14 +49,13 @@ const PaymentToolTip = () => (
 );
 
 const AgeRestrictionWarningMessage = () => (
-  <Message negative>
-    <p>Funktioniert nur in Verbindung mit einer MSAM Händler Karte</p>
+  <Message color="orange">
+    <p>Only works with MSAM dealer card.</p>
   </Message>
 );
 
 const CustomizeScreen = ({ cancelHandler, kioskProps, kiosk }) => {
   const dispatch = useDispatch();
-  console.log('Testing Kiosl', kiosk);
   const onSubmit = (values, formActions) => {
     const finalProps = {
       id: values.id,
@@ -86,6 +85,7 @@ const CustomizeScreen = ({ cancelHandler, kioskProps, kiosk }) => {
     setType(value);
     if (value === 'CreditOrDebitCard') {
       setAge('0');
+      setMemberCard(kioskProps.memberCardEnabled);
       setAgeRestrictionWarning(false);
     }
     if (value === 'GiroCard') {
@@ -107,12 +107,14 @@ const CustomizeScreen = ({ cancelHandler, kioskProps, kiosk }) => {
     if (kioskProps.minimumAge === '') setAge('0');
     else setAge(kioskProps.minimumAge.toString());
     if (!type || type === '') setType(kioskProps.paymentType);
+    setMemberCard(kioskProps.memberCardEnabled);
   }, []);
 
   useEffect(() => {
     if (kioskProps.minimumAge === '') setAge('0');
     else setAge(kioskProps.minimumAge.toString());
     setType(kioskProps.paymentType);
+    setMemberCard(kioskProps.memberCardEnabled);
     setServiceCheckEnabled(serviceCheckEnabled);
   }, [kioskProps]);
 
@@ -271,6 +273,9 @@ const CustomizeScreen = ({ cancelHandler, kioskProps, kiosk }) => {
                     isModalOpen={outOfServicewarning}
                     setIsModalOpen={setOutOfServiceWarning}
                     justConfirmation={true}
+                    onClickNo={() =>
+                      setFieldValue('serviceCheckEnabled', false)
+                    }
                   >
                     <p>
                       Are you sure you want to put this Kiosk ({kiosk.name}) in
@@ -309,9 +314,9 @@ const CustomizeScreen = ({ cancelHandler, kioskProps, kiosk }) => {
                   <MemberCardToolTip />
                   <Field
                     label="Enable"
-                    name="memberCardEnabled"
+                    name="memberCard"
                     value={true}
-                    checked={memberCard == true}
+                    checked={memberCard === true}
                     onChange={(e, { name, value }) => {
                       setMemberCard(value);
                       setFieldValue(name, value);
@@ -321,9 +326,9 @@ const CustomizeScreen = ({ cancelHandler, kioskProps, kiosk }) => {
                   />
                   <Field
                     label="Disable"
-                    name="memberCardDisabled"
+                    name="memberCard"
                     value={false}
-                    checked={memberCard == false}
+                    checked={memberCard === false}
                     onChange={(e, { name, value }) => {
                       setMemberCard(value);
                       setFieldValue(name, value);