+++ src/modules/kiosks/components/Toolbar.js
@@ -1,9 +1,18 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
-import { Segment, Grid, Input, Button, Icon, Dropdown, Divider} from 'semantic-ui-react';
+import {
+  Segment,
+  Grid,
+  Input,
+  Button,
+  Icon,
+  Dropdown,
+  Divider,
+} from 'semantic-ui-react';
 
-const Toolbar = ({ search,
+const Toolbar = ({
+  search,
   setSearch,
   kiosks,
   setKiosk,
@@ -17,11 +26,11 @@ const Toolbar = ({ search,
   };
   const handleKioskChange = (e, { value }) => {
     const text = value === 'All' ? '' : value;
-    setKiosk(text)
+    setKiosk(text);
   };
   const handleKiosksStatus = (e, { value }) => {
     const text = value === 'All' ? '' : value;
-    setKioskStatus(text)
+    setKioskStatus(text);
   };
   // const handleKiosksNetworkStatus = (e, { value }) => { //!LIV-2285
   //   const text = value === 'All' ? '' : value;
@@ -49,38 +58,38 @@ const Toolbar = ({ search,
               compact
               as={Link}
               // disabled={true} //! Disabled temporarily #golive2
-              to="/kiosks/edit/new" 
+              to="/kiosks/edit/new"
             >
               <Icon name="right arrow" />
               Add Kiosk
             </Button>
           </Grid.Column>
         </Grid.Row>
-      
-      <Divider style={{marginTop: 0, marginBottom: 0}} />
 
-      <Grid.Row verticalAlign="middle" columns="equal">
-        <Grid.Column width={3}>
-          <Dropdown
-              placeholder="All Kiosks"
+        <Divider style={{ marginTop: 0, marginBottom: 0 }} />
+
+        <Grid.Row verticalAlign="middle" columns="equal">
+          <Grid.Column width={3}>
+            <Dropdown
+              placeholder="Kiosks"
               selection
               className="full-width"
               onChange={handleKioskChange}
               options={kiosks}
-          />
-        </Grid.Column>
+            />
+          </Grid.Column>
 
-        <Grid.Column width={3}>
-          <Dropdown
+          <Grid.Column width={3}>
+            <Dropdown
               placeholder="Door Status"
               selection
               className="full-width"
               onChange={handleKiosksStatus}
               options={kiosksStatus}
-          />
-        </Grid.Column>
+            />
+          </Grid.Column>
 
-        {/* <Grid.Column width={3}> //!LIV-2285
+          {/* <Grid.Column width={3}> //!LIV-2285
           <Dropdown
               placeholder="Network Status"
               selection
@@ -89,8 +98,7 @@ const Toolbar = ({ search,
               options={kiosksNetworkStatus}
           />
         </Grid.Column> */}
-      </Grid.Row>
-      
+        </Grid.Row>
       </Grid>
     </Segment>
   );