+++ src/modules/shared/components/ConfirmationModal/index.js
@@ -8,13 +8,22 @@ const ConfirmationModal = ({
   title,
   children,
   justConfirmation,
+  onClickNo,
 }) => {
   return (
     <Modal size="mini" open={isModalOpen} onClose={() => setIsModalOpen(false)}>
       <Modal.Header>{title}</Modal.Header>
       <Modal.Content>{children}</Modal.Content>
       <Modal.Actions>
-        <Button negative onClick={() => setIsModalOpen(false)}>
+        <Button
+          negative
+          onClick={() => {
+            if (onClickNo) {
+              onClickNo();
+            }
+            setIsModalOpen(false);
+          }}
+        >
           No
         </Button>
         {justConfirmation ? (